% Auto-generated: do not edit by hand
\name{phydthreeComponent}

\alias{phydthreeComponent}

\title{PhydthreeComponent component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
phydthreeComponent(url=NULL)
}

\arguments{
\item{url}{Character. A label that will be printed when this component is rendered.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

